Add total dataset workflow
The original run.sh script is modified and moved to spydrpick_total folder,as well as src folder.
As the original version, this workflow runs the spydrpick analysis on the full sequence dataset, represented by
the multialignment provided by GISAID, filtering only the public sequences.
Sequences weights are re-calculated basing on tree distance and time, prior spydrpick analysis.
Only relevant MI interactions are conserved in final and intermediate outputs, labelled as outliers with 4 levels (1 to 4).

Old scripts post_filter.py and get_weights_linclust.py, which where made for the mmseqs clustering, are not part of the pipeline
anymore, but mantained in src/ just in case a cluster-based weight will be needed in future.